Phase 1: Area model DynamoDB implementation + full table schemas
- Fill in all 10 DynamoDB create-table definitions with complete
  attribute definitions, key schemas, and GSIs in 00-create-dynamodb-tables.sh.
- Add area.dynamo.js: DynamoDB-backed area model supporting get, findById,
  findOne({year}), save, deleteOne. Admin aggregation paths stubbed with
  NotImplementedError (out of scope per plan).
- Convert area.model.js to a feature-flag switch file: routes to
  area.model.mongoose.js (original) or area.dynamo.js based on
  config.dynamodb.useAreas. Transparent to controllers.
- 172 tests passing, zero regressions.
